Move unit tests to the new place
